+++ src/FrbaOfertas/FrbaOfertas/Models/Facturas/ItemFactura.cs
@@ -12,7 +12,7 @@ namespace FrbaOfertas.Models.Facturas
         public int id_item_factura { get; set; }
         public Factura factura { get; set; }
         public Oferta oferta { get; set; }
-        public int importe_oferta { get; set; }
+        public decimal importe_oferta { get; set; }
         public int cantidad { get; set; }
 
         public ItemFactura() { }
